AIServer/requirements.txt
fastapi
Flask==2.0.1
gunicorn==20.1.0
urllib.parse
langchain-openai
openai
requests
pinecone
python-dotenv

